+++ example/example.js
@@ -22,6 +22,7 @@ const k = {
   address: '0.0.0.0',
   port: 3000,
   baseNamespace: 'com.github.smartiniOnGitHub.fastify-cloudevents.example',
+  includeHeaders: false, // same as default value
   cloudEventOptions: {
     strict: true // enable strict mode in generated CloudEvents, optional
   }
@@ -46,6 +47,7 @@ const gen = idCounterExample()
 fastify.register(require('../src/plugin'), {
   serverUrl: k.serverUrl,
   idGenerator: gen,
+  includeHeaders: k.includeHeaders,
   onRequestCallback: loggingCallback,
   preHandlerCallback: loggingCallback,
   onSendCallback: loggingCallback,